pyproject.toml
[project]
name = "ai-bi"
version = "0.1.0"
description = "AI-nalyst com LangGraph + dbt-DuckDB + Streamlit"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "streamlit>=1.43.0",
    "pandas>=2.2.0",
    "duckdb>=1.2.0",
    "pyarrow>=19.0.0",
    "plotly>=6.0.0",
    "python-dotenv>=1.1.0",
    "kagglehub>=0.3.0",
    "dbt-duckdb>=1.9.0",
    "langchain-openai>=0.3.0",
    "langchain-core>=0.3.0",
    "langgraph>=0.3.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
